src/requirements.txt
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
decorator==5.1.1
Django==5.1.3
django-browser-reload==1.17.0
django-extensions==3.2.3
django-simple-history==3.7.0
django-tailwind==3.8.0
executing==2.1.0
idna==3.10
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
parso==0.8.4
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
